How to convert this Elvui skins to original Blizzard skins
if you don't want the chat appearance like this and prefer the original Blizzard one, you need to go to the ElvUI settings, navigate to chat settings, and uncheck the "enable" button at the top to adjust chat settings for ElvUI. You can do the same for action bars or bags. Lastly, you can revert menus like achievements or group finder to Blizzard's original. To do this, go to the ElvUI settings, find the "skins" section, and uncheck the ones you don't want.